(SIMP-1736) Hiera-based parameter enforcement
This implements a Hiera backend that has the ability to stack various
compliance profiles in such a way that a unified parameter stack can be
applied successfully across *multiple* compliance data sets.

Presently, it is using a recursive lookup in the backend itself but has
a recursion lock to prevent insanity from ensuing.

When the PE LTS release is gone, and Hiera 5 is fully in place, we
should migrate this to a data funcion. But, for the duration of SIMP 6,
this is what we've got so that we can have the capability.

This should be pulled in and used in acceptance testing across *ALL*
SIMP modules since it provides for rapid implementation testing of the
various components and expansion of the back-end data sets.

SIMP-1736 #close